move compiler code
